Scoping: any file_based specific implementation is now only used in remote_factory, utils, and manager.file_based
