+++ pacco/cli/binary.py
@@ -7,7 +7,7 @@ import shutil
 from pathlib import Path
 from typing import Dict, Callable
 
-from pacco.cache.cache import Cache
+from pacco.manager.utils.cache import Cache
 from pacco.cli import utils
 from pacco.manager.remote_manager import RemoteManager
